1)Se implementa documentación de las respuestas de los endpoints con HATEOAS. 2) Se genera capa de pruebas unitarias.

+++ src/main/java/com/crud_peliculas/repository/PeliculaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.crud_peliculas.model.entities.Pelicula;
 
-public interface PeliculaRepository extends JpaRepository<Pelicula, Integer>{
+public interface PeliculaRepository extends JpaRepository<Pelicula, Long>{
 
 }
